# alternator: eliminate regexes from the hot path