@nuivall 2023-01-04
> By the way, I had a thought: Why must table name validation be in the hot path?
> 
> Surely, we need to validate table names in the `CreateTable` request. But why do we need to do it in other requests like GetItem and so on? Do they really need to bother with checking the "validity" of the table name? Can't we just check if the table exists, and if it doesn't exist then it doesn't really matter if it doesn't exist because it wasn't created, or because it couldn't be created because it has a bad name?
> 
> Or maybe we want to do this checking for nicer or more-compatible error handling, so trying to read from table "!@#" returns an error about an invalid name instead of just an error that the table doesn't exit?

Yes I was thinking about it too. Why not just rely on some kind of `not found` from lower layer. Although it would change some logic which I don't fully understand and I wanted to limit the scope of the patch to just low hanging fruit.