@nuivall 2023-01-04
> I'm a bit surprised that giving up on regular expressions improves performance. I always wanted to pretend that once a regular expression is compiled, a lot of effort and brains went into make sure that executing the compiled expression is the fastest it could be, using clever lookup tables and modern simd instructions or who knows what to make matching this trivial expression very fast. But I guess std::regex doesn't really have so much cleverness and a trivial hand-written loop with no optimization can do noticably better :-(

Empirical reasons for it being slow in this case seems to be that implementation creates "high" stacks, another is that it allocates memory (haven't checked where). And my guess is that simd doesn't help much if string is just couple chars long, it's loaded at once anyway and then read from cache. And if it is faster it's not noticable because even after this patch the function disappeared from the profile completely.

> Turns out that std::regex is "notorious" for bad performance - see for example https://stackoverflow.com/questions/70583395/why-is-stdregex-notoriously-much-slower-than-other-regular-expression-librarie ...

Yes, after seeing how big difference it makes I started to wonder what opportunity lies in other 76 regexes in the codebase :)