@avikivity 2023-01-05
Nice improvement, but the overall numbers are frightening. Why did you base your protocol on http/json?!

Possible areas of improvement are to change the parser to work with views rather than allocate strings. Of course, that's a monumental task, especially if the input is fragmented.