@avikivity 2023-01-05
/cc @michoecho makes CQL looks good