@michoecho 2023-01-05
@nyh @nuivall @avikivity Didn't see anyone post one, so here is a Alternator vs CQL simple query (cache hit) trace comparison (same table, same partition, both time axes with a common scale):

![image](https://user-images.githubusercontent.com/45052050/210793116-ab3c9d45-b711-419d-94d7-3b8ce7a883f4.png)
 
Here are the individual traces for close viewing (you can open them and scroll/zoom in https://ui.perfetto.dev/). 

[alternator_vs_cql.zip](https://github.com/scylladb/scylladb/files/10352524/alternator_vs_cql.zip)

The table is the same as used here: https://github.com/awsdocs/aws-doc-sdk-examples/blob/main/python/example_code/dynamodb/GettingStarted/scenario_getting_started_movies.py