@michoecho 2023-01-05
Alternator is ~50% parsing, so it's 2x slower than CQL in this case.