@avikivity 2023-01-05
DynamoDB has an almost-prepared-statement thing: https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_Update.html, see ExpressionAttribute{Name,Value}s. Do we know if the SDKs use these, or if they inline values into the json body?

If SDKs consistently use them, we can try to extract prepared statements from the input json, and then we only have to parse the variable part of the json (ExpressionAttributeValues). We convert the input json into a constant prefix + constant suffix, and then only have to parse the variable part (and perhaps even not all of it).