@avikivity 2023-01-05
> Alternator is ~50% parsing, so it's 2x slower than CQL in this case.

btw, it will be diluted for real workloads that double the cache work (for QUORUM reads) and add some rpc overhead.