@michoecho 2023-01-05
> btw, it will be diluted for real workloads that double the cache work (for QUORUM reads) and add some rpc overhead.

It will also be un-diluted in a real workload by the fact that non-parsing will get compressed by execution-stage batching, but parsing will not, because parsing is a tighter loop. But...

Look at the trace, 300000 instructions is spent in the _http_ request parser. Does that seem reasonable? Maybe something is wrong with this parser.