@avikivity 2023-01-05
> 
> Look at the trace, 300000 instructions is spent in the _http_ request parser. Does that seem reasonable? Maybe something is wrong with this parser.

That code is maybe 8 years old and was written to get a quick proof that Seastar works.

But it actually should be good, it's using a regular expression description language. It's not parsing the body, just the headers. Odd.