@michoecho 2023-01-05
About json, from the trace it seems that the impact of json is actually very small. The meat of the overhead over CQL seems to be in the ragel http parser and in the antlr expression parser in alternator.