@avikivity 2023-01-05
> > About json, from the trace it seems that the impact of json is actually very small. The meat of the overhead over CQL seems to be in the ragel http parser and in the antlr expression parser in alternator.
> 
> That's good news. If the SDK uses references in expressions, then the expression strings will be reused across calls, and we can cache the parse results.

(your example expression fits, but we have to see if popular SDKs also do that).