@michoecho 2023-01-05
> (your example expression fits, but we have to see if popular SDKs also do that).

I used boto3 for this (I mostly copy-pasted code from the AWS docs example I linked above), so Python is covered.