@michoecho 2023-01-05
But again, I would start any performance investigations here with the HTTP parser, it takes 25% of coordinator's work (most likely more in a real workload, due to hyperthreading and cache effects), and that just doesn't feel right.